+++ lib/src/parse/keyframe_selector.dart
@@ -4,12 +4,14 @@
 
 import 'package:charcode/charcode.dart';
 
+import '../logger.dart';
 import '../util/character.dart';
 import 'parser.dart';
 
 /// A parser for `@keyframes` block selectors.
 class KeyframeSelectorParser extends Parser {
-  KeyframeSelectorParser(String contents, {url}) : super(contents, url: url);
+  KeyframeSelectorParser(String contents, {url, Logger logger})
+      : super(contents, url: url, logger: logger);
 
   List<String> parse() {
     return wrapSpanFormatException(() {